# aspnetcore/src/IO.Swagger/Controllers/DefaultApi.cs
/*
 * defaultTitle
 *
 * defaultDescription
 *
 * OpenAPI spec version: 0.1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class DefaultApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Open API Specification for Enrichment Event Rest</remarks>
        /// <param name="body"></param>
        /// <response code="200">Response for Enrichment Event</response>
        [HttpPost]
        [Route("/life-cycle-service/v1/enrichmentEvents/addAll")]
        [ValidateModelState]
        [SwaggerOperation("LifeCycleServiceV1EnrichmentEventsAddAllPost")]
        [SwaggerResponse(statusCode: 200, type: typeof(InlineResponse200), description: "Response for Enrichment Event")]
        public virtual IActionResult LifeCycleServiceV1EnrichmentEventsAddAllPost([FromBody]Body body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(InlineResponse200));
            string exampleJson = null;
            exampleJson = "{\n  \"errorList\" : [ \"errorList\", \"errorList\" ],\n  \"message\" : \"message\"\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<InlineResponse200>(exampleJson)
                        : default(InlineResponse200);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}